+++ Plugins/PluginPdf/pluginpdf.cpp
@@ -5,7 +5,7 @@ PluginPdf::PluginPdf(QObject *parent) : QObject(parent)
     m_wMainWindow = new MainWindow;
 }
 
-void PluginPdf::sendFileData(char *szFileData, int nFileLen)
+void PluginPdf::sendFileData(char *szFileData, qint64 nFileLen)
 {
     m_wMainWindow->showPdf(szFileData, nFileLen);
 }
