+++ Plugins/PluginText/plugintext.h
@@ -12,7 +12,7 @@ class PluginText : public QObject,public PluginInterface
     Q_PLUGIN_METADATA(IID "fileWrapper.pluginText" FILE "pluginText.json")
 public:
     explicit PluginText(QObject *parent = nullptr);
-    void sendFileData(char *szFileData, int nFileLen);
+    void sendFileData(char *szFileData, qint64 nFileLen);
     QWidget* getPluginWidget();
 
 private:
